Add clearAll.js file
