+++ src/index.js
@@ -12,6 +12,9 @@ import removeFromStorage from './modules/removeFromStorage.js';
 
 import toggleCompleted from './modules/toggleCompleted.js';
 
+import clearAll from './modules/clearAll.js';
+
+
 document.addEventListener('DOMContentLoaded', Interface.displayTasks);
 
 document.querySelector('.form').addEventListener('submit', (e) => {
@@ -61,13 +64,4 @@ document.querySelector('#toDoItems').addEventListener('keypress', (e) => {
 
 document.querySelector('#toDoItems').addEventListener('change', toggleCompleted);
 
-document.querySelector('#clearAllBtn').addEventListener('click', () => {
-  let tasks = [];
-  if (localStorage.getItem('tasks') !== null) {
-    tasks = JSON.parse(localStorage.getItem('tasks'));
-  }
-  const newTasks = tasks.filter((task) => task.completed === false);
-  localStorage.setItem('tasks', JSON.stringify(newTasks));
-  updateid.storage();
-  Interface.displayTasks();
-});
+document.querySelector('#clearAllBtn').addEventListener('click', clearAll);
