+++ src/modules/tests/toggleCompleted.test.js
@@ -1,15 +1,8 @@
 /** * @jest-environment jsdom */
 
-import toggleCompleted from "../toggleCompleted.js";
-import Storage from "../storage.js";
+import toggleCompleted from '../toggleCompleted.js';
 
 jest.mock('../toggleCompleted.js');
-
-beforeEach(() => {
-  toggleCompleted.mockClear();
-});
-
-window.localStorage = Storage.prototype;
 
 const tasks = [
   {
@@ -24,7 +17,8 @@ localStorage.setItem('tasks', JSON.stringify(tasks));
 
 describe('Updating completed status', () => {
   test('if the completed attribute changes', () => {
-    toggleCompleted(0);
-    expect(tasks[0].completed).toBeTruthy();
+    toggleCompleted();
+    const dataStorage = JSON.parse(localStorage.getItem('tasks')) || [];
+    expect(dataStorage[0].completed).toBeTruthy();
   });
 });
